ml-service/requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
fastapi==0.111.1
uvicorn==0.30.6
python-dotenv==1.0.1
matplotlib==3.9.4
scipy==1.13.1
azure-storage-blob==12.23.1
